requirements.txt
langgraph
langchain_openai
langchain_teddynote
faiss-cpu
pdfplumber
langchain_community